refactoring part 1

+++ src/citra_libretro/input/mouse_tracker.h
@@ -19,7 +19,7 @@ namespace Input {
 class CursorRenderer {
 public:
     virtual ~CursorRenderer() = default;
-    virtual void Render(int bufferWidth, int bufferHeight, float projectedX, float projectedY,
+    virtual void Render(int bufferWidth, int bufferHeight, int projectedX, int projectedY,
                         float renderRatio, const Layout::FramebufferLayout& layout,
                         void* framebuffer_data = nullptr) = 0;
 };
@@ -32,10 +32,10 @@ public:
     ~MouseTracker();
 
     /// Called whenever a mouse moves.
-    void OnMouseMove(int xDelta, int yDelta);
+    void OnMouseMove(float xDelta, float yDelta);
 
     /// Restricts the mouse cursor to a specified rectangle.
-    void Restrict(int minX, int minY, int maxX, int maxY);
+    void Restrict(float minX, float minY, float maxX, float maxY);
 
     /// Updates the tracker.
     void Update(int bufferWidth, int bufferHeight, const Layout::FramebufferLayout& layout);
@@ -60,8 +60,8 @@ public:
     }
 
 private:
-    int x;
-    int y;
+    float x;
+    float y;
 
     float lastMouseX;
     float lastMouseY;
@@ -80,7 +80,7 @@ class OpenGLCursorRenderer : public CursorRenderer {
 public:
     OpenGLCursorRenderer();
     ~OpenGLCursorRenderer();
-    void Render(int bufferWidth, int bufferHeight, float projectedX, float projectedY,
+    void Render(int bufferWidth, int bufferHeight, int projectedX, int projectedY,
                 float renderRatio, const Layout::FramebufferLayout& layout,
                 void* framebuffer_data = nullptr) override;
 
@@ -96,7 +96,7 @@ class VulkanCursorRenderer : public CursorRenderer {
 public:
     VulkanCursorRenderer();
     ~VulkanCursorRenderer();
-    void Render(int bufferWidth, int bufferHeight, float projectedX, float projectedY,
+    void Render(int bufferWidth, int bufferHeight, int projectedX, int projectedY,
                 float renderRatio, const Layout::FramebufferLayout& layout,
                 void* framebuffer_data = nullptr) override;
 };
@@ -106,7 +106,7 @@ class SoftwareCursorRenderer : public CursorRenderer {
 public:
     SoftwareCursorRenderer();
     ~SoftwareCursorRenderer();
-    void Render(int bufferWidth, int bufferHeight, float projectedX, float projectedY,
+    void Render(int bufferWidth, int bufferHeight, int projectedX, int projectedY,
                 float renderRatio, const Layout::FramebufferLayout& layout,
                 void* framebuffer_data = nullptr) override;
 };
